Revert "Merge branch 'develop' into feat/P4ADEV-3355-assessment-list-page"
This reverts commit 4b06b4c8030c7492ce998efaf693cff470c34add, reversing
changes made to 1023b0cdc0996357f85fcb622d06af6fcb1162b6.
