+++ src/api/getPagoPaRegistry.ts
@@ -14,7 +14,7 @@ const getPagoPaRegistries = (organizationId: OrgID) =>
         organizationId,
         query
       );
-      parseAndLog(pagedPagoPaRegistrySchema, response, false);
+      parseAndLog(pagedPagoPaRegistrySchema, response);
       return response;
     }
   });
